Rename MiddlewareMacro
