+++ templates/cocos2dx_files.json
@@ -153,7 +153,7 @@
         "cocos/editor-support/IOBuffer.h", 
         "cocos/editor-support/IOTypedArray.cpp", 
         "cocos/editor-support/IOTypedArray.h", 
-        "cocos/editor-support/MiddleWareMacro.h", 
+        "cocos/editor-support/MiddlewareMacro.h", 
         "cocos/editor-support/MiddlewareManager.cpp", 
         "cocos/editor-support/MiddlewareManager.h", 
         "cocos/editor-support/TypedArrayPool.cpp",
